add nats

+++ pkg/tagdata/model.go
@@ -3,11 +3,11 @@ package tagdata
 type TagData struct {
 	Name    string    `json:"name"`
 	Dscr    string    `json:"dscr"`
-	Min     float64   `json:"min"`
-	Max     float64   `json:"max"`
+	Min     float32   `json:"min"`
+	Max     float32   `json:"max"`
 	CycleMs int       `json:"cyclems"`
 	Unit    string    `json:"senstype"`
-	V       []float64 `json:"v"`
+	V       []float32 `json:"v"`
 	//Y       []opts.LineData `json:"-"`
 	//T       []string        `json:"t"`
 }
